requirements.txt
# AgroConnect API — Python 3.11+
fastapi>=0.115.0,<0.116
uvicorn[standard]>=0.32.0,<0.33
asyncpg>=0.30.0,<0.31
httpx>=0.28.0,<0.29
pydantic[email]>=2.10.0,<3
python-jose[cryptography]>=3.3.0,<4
passlib[bcrypt]>=1.7.4,<2
# bcrypt 4.1+ supprime __about__ — incompatible avec passlib 1.7.x sur certains setups
bcrypt>=4.0.1,<4.1
python-multipart>=0.0.12,<0.1
email-validator>=2.2.0,<3
apscheduler>=3.10.4,<4
python-dotenv>=1.0.0,<2
